requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
autopep8==2.3.2
black==25.1.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.8
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.67.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
